Add example use of DatabaseLibrary
This is a library provided by the instructor for CS 230.  At this point,
we are unable to change this library for our project, but we'll need
to use it to access, store, and update data for the CMC system.

Currently, the library works but anyone running it needs to add "Arguments"
to the run in their Eclipse project.  To do so, just right-click and go to
"Run Configurations" when trying to Run this Java program in Eclipse.